backend/requirements/dev.txt
-r base.txt

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Development
django-debug-toolbar>=4.2,<5.0
ipython>=8.0,<9.0

# Code quality
flake8>=7.0,<8.0
black>=24.0,<25.0
isort>=5.13,<6.0
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0

# Pre-commit
pre-commit>=3.6,<4.0
